change EAD schema to use xlink.xsd from DLTS server
Change the xlink.xsd schemaLocation URL in the EAD 2002 schema from a
Library of Congress (LoC) URL (http://www.loc.gov/standards/xlink/xlink.xsd)
to a URL on a DLTS server
(http://ead-publisher.dlib.nyu.edu/support/xsd/xlink.xsd).

This change was made because the tests and EAD validation
functionality started failing after a prolonged periood of testing and
validation operations. The libxml2 error messages indicated that the
xlink.xsd schema could not be retrieved from the LoC server.  Further
investigation revealed that the LoC server was returning a 403
Forbidden status.  I believe that a rate-limiting mechanism was
triggered by the prolonged period of testing and validation
operations.

To avoid this rate-limiting behavior, a copy of the xlink schema was
placed on a DLTS server and the schemaLocation in the EAD 2002 schema
in this repo was updated to use the DLTS URL.